fix updated according to the comments made
