+++ frontend/src/app/routes/admin-page/admin-page.scss
@@ -18,20 +18,13 @@
 
 }
 
-/* .admin-tool-bar {
-  flex: 0 0 auto;
-} */
 
 .admin-container {
-  /*   width: 100%;
-  height: 100%; */
   flex: 1;
   padding: 24px;
   display: flex;
   flex-direction: column;
   gap: 10px;
-  //min-height: 0;
-  //background-color: --var(mat-sys-surface-container);
   overflow: auto;
   margin: 0 auto; //centrerat i mitten
